split instrument calibration again for now and more small linter fixes
