+++ api/src/opentrons/hardware_control/instruments/ot3/instrument_calibration.py
@@ -15,6 +15,7 @@ from opentrons.calibration_storage import (
     get_pipette_offset,
     ot3_gripper_offset,
 )
+from opentrons.calibration_storage.ot3 import models as ot3_models
 from opentrons.hardware_control.types import OT3Mount
 
 
@@ -61,7 +62,10 @@ def load_pipette_offset(
     else:
         checked_mount = mount
     if pip_id:
-        pip_offset_data = get_pipette_offset(pip_id, checked_mount)
+        pip_offset_data = typing.cast(
+            ot3_models.v1.InstrumentOffsetModel,
+            get_pipette_offset(pip_id, checked_mount),
+        )
         if pip_offset_data:
             return PipetteOffsetByPipetteMount(
                 offset=pip_offset_data.offset,
@@ -86,7 +90,7 @@ def save_pipette_offset_calibration(
     else:
         checked_mount = mount
     if pip_id:
-        save_pipette_calibration(offset, pip_id, checked_mount)
+        save_pipette_calibration(offset, pip_id, checked_mount, calibration_status=None)
 
 
 def load_gripper_calibration_offset(
