+++ api/tests/opentrons/calibration_storage/__init__.py
@@ -1,5 +1,4 @@
 import typing
-from json import JSONDecoder
 from pydantic import BaseModel
 from pathlib import Path
